Rename WalletViewNav -> AssetsNav + add a story
for it
