+++ src/components/wallet/AssetsNav.stories.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import { storiesOf } from '@storybook/react'
+
+import AssetsNav from './AssetsNav'
+
+storiesOf('Wallet/AssetsNav', module).add('default', () => {
+  return <AssetsNav />
+})
